+++ js/routes/pantry-queue.js
@@ -1,28 +1,62 @@
 import { html, el } from '@aegisjsproject/core/parsers/html.js';
 import { data, attr } from '@aegisjsproject/core/stringify.js';
 import { registerCallback } from '@aegisjsproject/callback-registry/callbacks.js';
-import { onClick, onSubmit, onReset, onChange, onClose, signal as signalAttr, registerSignal } from '@aegisjsproject/callback-registry/events.js';
+import { onClick, onSubmit, onReset, onClose, signal as signalAttr, registerSignal } from '@aegisjsproject/callback-registry/events.js';
 import { openDB, getItem, getAllItems, deleteItem, putItem } from '@aegisjsproject/idb';
+import { whenLoaded } from '@aegisjsproject/router';
 import { SCHEMA } from '../consts.js';
 import { createBarcodeScanner, preloadRxing, QR_CODE } from '@aegisjsproject/barcodescanner';
-import { fetchWellKnownKey } from '@shgysk8zer0/jwk-utils/jwk.js';
-import { verifyJWT } from '@shgysk8zer0/jwk-utils/jwt.js';
 import { createSVGElement } from '@aegisjsproject/qr-encoder';
-import { TOWNS, ZIPS, postalCodes } from './pantry.js';
-import { HOUSEHOLD_LIST, getPantryHouseholdTemplate, pantryAddHousehold, HOUSEHOLD_MEMBER_CLASSNAME } from '../components/pantry.js';
+import { getPantryHouseholdTemplate, HOUSEHOLD_MEMBER_CLASSNAME } from '../components/pantry.js';
 
+const PTS = [
+	30, // 1
+	60, // 2
+	80, // 3
+	95, // 4
+	110, // 5
+	120, // 6
+	125, // 7
+	130, // 8
+	135, // 9
+	140, //10
+];
+
+const MAX_HOUSEHOLD = PTS.length;
+const BASE_POINTS = 5;
 const ID = 'pantry-queue';
 const STORE_NAME = 'pantryQueue';
 const ADD_FORM_ID = 'pantry-queue-form';
 const ADD_DIALOG_ID = 'pantry-queue-modal';
-
-const key = await fetchWellKnownKey(location.origin);
+const postalCodes = {
+	'alta sierra': '95949',
+	'weldon': '93283',
+	'bodfish': '93205',
+	'south lake': '93240',
+	'mt mesa': '93240',
+	'mountain mesa': '93240',
+	'wofford heights': '93285',
+	'lake isabella': '93240',
+	'kernville': '93238',
+	'onyx': '93255',
+	'canebrake': '93255',
+	'havilah': '93518',
+	'caliente': '93518',
+	'squirrel mountain valley': '93240',
+	'squirrel valley': '93240',
+	'keyesville': '93240',
+	'keysville': '93240',
+};
 
 const closeModal = registerCallback('pantry:queue:close-modal', ({ currentTarget }) => currentTarget.closest('dialog').close());
 
+function _getPoints(household) {
+	return PTS[Math.min(Math.max(parseInt(household), 1), MAX_HOUSEHOLD) - 1];
+}
+
 const closeAndRemove = registerCallback('pantry:queue:close-and-remove', async ({ currentTarget }) => {
 	await removeVisit(currentTarget);
-	document.getElementById(`visit-${currentTarget.dataset.txn}`).remove();
+	document.getElementById(`visit-${currentTarget.dataset.id}`).remove();
 	currentTarget.closest('dialog').close();
 });
 
@@ -34,44 +68,49 @@ export const updateZip = registerCallback('pantry:queue:zip-update', ({ target: 
 	}
 });
 
+/**
+ * @todo Rewrite to generate JWT here instead of POSTing
+ */
 const submitHandler = registerCallback('pantry:queue:submit', async event => {
 	event.preventDefault();
 	// Store the submitter, with a default empty object just in case.
-	const submitter = event.submitter ?? {};
+	const { submitter = {}, target } = event;
+	const db = await _openDB();
 
 	try {
 		submitter.disabled = true;
-		/**
-		 * @type HTMLFormElement
-		 */
-		const target = event.target;
 		const data = new FormData(target);
-		data.set('datetime', new Date(data.get('date') + 'T' + data.get('time')).toISOString());
+		const date = new Date();
+		const id = '_' + crypto.randomUUID();
+		const url = new URL('/pantry/distribution', location.origin);
+		const params = new URLSearchParams(data);
+		const name = ['givenName', 'additionalName', 'familyName', 'suffix']
+			.map(field => data.has(field) ? data.get(field).trim() : null)
+			.filter(field => typeof field === 'string' && field.length !== 0)
+			.join(' ');
 
-		const resp = await fetch('/api/pantry', {
-			method: 'POST',
-			body: data,
-		});
+		const household = parseInt(data.get('household'));
+		const normalTrip = ! data.has('extraTrip');
+		const points = normalTrip ? _getPoints(household) : Math.min(Math.max(household, 1), PTS.length) * BASE_POINTS;
+		params.set('name', name);
+		params.set('time', date.toISOString());
+		params.set('points', points);
+		params.set('household', household);
+		params.set('id', id);
+		url.search = params;
 
-		if (resp.ok) {
-			const body = await resp.formData();
-			const jwt = body.get('jwt');
-			await checkInVisit({ rawValue: jwt });
-			event.target.reset();
-			submitter.disabled = false;
+		await putItem(db, STORE_NAME, { sub: name, txn: id, household, points, date });
 
-			if (submitter?.dataset?.close === 'true') {
-				document.getElementById(ADD_DIALOG_ID).requestClose();
-			} else {
-				target.scrollIntoView({ behavior: 'smooth', block: 'start' });
-			}
-		} else {
-			const err = await resp.json();
-			throw new Error(err.error.message);
+		if (typeof submitter.dataset.close === 'string') {
+			target.closest('dialog')?.requestClose?.();
 		}
+		await render();
+		target.reset();
 	} catch(err) {
 		alert(err.message);
+	} finally {
 		submitter.disabled = false;
+		db.close();
 	}
 });
 
@@ -95,18 +134,26 @@ async function showVisit(btn) {
 		const visit = await getVisit(btn.dataset.txn);
 
 		if (typeof visit === 'object') {
+			const url = new URL('/pantry/distribution', location.origin);
+			const params = new URLSearchParams();
+			params.set('name', visit.sub);
+			params.set('id', visit.txn);
+			params.set('household', visit.household);
+			params.set('points', visit.points);
+			params.set('date', visit.date.toISOString());
+			url.search = params;
 			const dialog = el`<dialog id="modal-${visit.txn}" ${onClose}="${closeHandler}">
 				<p>Visit scheduled for ${visit.sub} at <time ${attr({ datetime: visit.date.toISOString() })}>${visit.date.toLocaleTimeString()}</p>
-				${createSVGElement(visit.jwt, { size: 480 })}
+				${createSVGElement(url.href, { size: 480 })}
 				<hr />
 				<button type="button" class="btn btn-warning" ${onClick}="${closeModal}">Close</button>
-				<button type="button" class="btn btn-danger" ${onClick}="${closeAndRemove}" ${data({ txn: visit.txn })}>Close &amp; Remove</button>
+				<button type="button" class="btn btn-danger" ${onClick}="${closeAndRemove}" ${data({ id: visit.txn })}>Close &amp; Remove</button>
 			</dialog>`;
 
 			document.getElementById('main').append(dialog);
 			dialog.showModal();
 		} else {
-			throw new Error(`No results for ${btn.dataset.txn}.`);
+			throw new Error(`No results for ${btn.dataset.id}.`);
 		}
 	} catch(err) {
 		reportError(err);
@@ -122,7 +169,7 @@ async function removeVisit(btn) {
 
 	try {
 		const row = btn.closest('tr');
-		await deleteItem(db, STORE_NAME, btn.dataset.txn);
+		await deleteItem(db, STORE_NAME, btn.dataset.id ?? btn.dataset.txn);
 		row.remove();
 	} catch(err) {
 		reportError(err);
@@ -146,11 +193,11 @@ const clickHandler = registerCallback('pantry:queue:show', async event => {
 	}
 });
 
-async function getVisit(txn) {
+async function getVisit(id) {
 	const db = await _openDB();
 
 	try {
-		const visit = await getItem(db, STORE_NAME, txn);
+		const visit = await getItem(db, STORE_NAME, id);
 		db.close();
 		return visit;
 	} catch (err) {
@@ -164,7 +211,7 @@ async function render() {
 
 	try {
 		const visits = await getAllItems(db, STORE_NAME);
-		const rows = visits.sort(_sort).map(v => createVisitRow(v, { parse: true }));
+		const rows = visits.sort(_sort).map(({ sub: name, txn: id, household, points, date}) => createVisitRow({ name, id, household, points, date}, { parse: true }));
 
 		document.getElementById(ID).tBodies.item(0).replaceChildren(...rows);
 	} finally {
@@ -173,8 +220,6 @@ async function render() {
 }
 
 async function checkInVisit({ rawValue }) {
-	const now = Math.floor(Date.now() / 1000);
-
 	if (URL.canParse(rawValue)) {
 		const url = new URL(rawValue);
 
@@ -195,64 +240,40 @@ async function checkInVisit({ rawValue }) {
 
 			modal.showModal();
 		}
-	} else {
-		const {
-			sub,
-			toe,
-			txn,
-			authorization_details: { household, points },
-		} = await verifyJWT(rawValue, key);
-
-		const db = await _openDB();
-
-		try {
-			const date = now > toe ? new Date() : new Date(toe * 1000);
-			await putItem(db, STORE_NAME, { sub, txn, household, points, date, checkedIn: new Date(), jwt: rawValue });
-			await render();
-		} finally {
-			db.close();
-		}
 	}
-
 }
 
 preloadRxing();
 
-function createVisitRow({ sub, txn, household, date, checkedIn }, { parse = false } = {}) {
+function createVisitRow({ name, id, household, date }, { parse = false } = {}) {
 	if (parse) {
-		return html`<tr id="visit-${txn}" ${data({ txn })}>
-			<td class="visit-name">${sub}</td>
+		return html`<tr id="visit-${id}" ${data({ id })}>
+			<td class="visit-name">${name}</td>
 			<td class="visit-time">
 				<time ${attr({ datetime: date.toISOString() })}>${date.toLocaleTimeString()}</time>
 			</td>
-			<td class="visit-check-in">
-				<time ${attr({ datetime: checkedIn.toISOString() })}>${checkedIn.toLocaleTimeString()}</time>
-			</td>
 			<td>${household}</td>
 			<th>
-				<button type="button" class="btn btn-primary" ${data({ txn })} data-queue-action="show">Show QR</button>
-				<button type="button" class="btn btn-danger" ${data({ txn })} data-queue-action="remove">Remove</button>
+				<button type="button" class="btn btn-primary" ${data({ txn: id })} data-queue-action="show">Show QR</button>
+				<button type="button" class="btn btn-danger" ${data({ txn: id })} data-queue-action="remove">Remove</button>
 			</th>
 		</tr>`;
 	} else {
-		return `<tr id="visit-${txn}" ${data({ txn })}>
-			<td class="visit-name">${sub}</td>
+		return `<tr id="visit-${id}" ${data({ id })}>
+			<td class="visit-name">${name}</td>
 			<td class="visit-time">
 				<time ${attr({ datetime: date.toISOString() })}>${date.toLocaleTimeString()}</time>
 			</td>
-			<td class="visit-check-in">
-				<time ${attr({ datetime: checkedIn.toISOString() })}>${checkedIn.toLocaleTimeString()}</time>
-			</td>
 			<td>${household}</td>
 			<th>
-				<button type="button" class="btn btn-primary" ${data({ txn })} data-queue-action="show">Show QR</button>
-				<button type="button" class="btn btn-danger" ${data({ txn })} data-queue-action="remove">Remove</button>
+				<button type="button" class="btn btn-primary" ${data({ txn: id })} data-queue-action="show">Show QR</button>
+				<button type="button" class="btn btn-danger" ${data({ txn: id })} data-queue-action="remove">Remove</button>
 			</th>
 		</tr>`;
 	}
 }
 
-export default async ({ signal: sig }) => {
+export default async ({ signal: sig, url }) => {
 	const signal = registerSignal(sig);
 
 	const { video } = await createBarcodeScanner(checkInVisit, { signal: sig, formats: [QR_CODE] });
@@ -268,12 +289,11 @@ export default async ({ signal: sig }) => {
 		<thead>
 			<th>Name</th>
 			<th>Time</th>
-			<th>Check-In</th>
 			<th>Household</th>
 			<th>Actions</th>
 		</thead>
 		<tbody id="${ID}-body">
-			${visits.sort(_sort).map(createVisitRow).join('')}
+			${visits.sort(_sort).map(({ sub: name, txn: id, household, points, date }) => createVisitRow({ name, id, household, points, date })).join('')}
 		</tbody>
 	</table>
 	<dialog id="${ADD_DIALOG_ID}">
@@ -313,52 +333,12 @@ export default async ({ signal: sig }) => {
 					</span>
 				</div>
 				<div class="form-group">
-					<label for="pantry-bday" class="input-label required">Birthdate</label>
-					<input type="date" name="bDay" id="pantry-bday" class="input" autocomplete="bday" placeholder="YYYY-MM-DD *" required="" />
+					<label for="${ADD_FORM_ID}-household" class="input-label required">Household Size</label>
+					<input type="number" name="household" id="${ADD_FORM_ID}-household" class="input" placeholder="##" min="1" max="${MAX_HOUSEHOLD}" required="" />
 				</div>
-				<div class="form-group">
-					<label for="${ADD_FORM_ID}-email" class="input-label">Email</label>
-					<input type="email" name="email" id="${ADD_FORM_ID}-email" class="input" placeholder="user@example.com" />
-				</div>
-				<div class="form-group">
-					<label for="${ADD_FORM_ID}-phone" class="input-label">Phone</label>
-					<input type="tel" name="telephone" id="${ADD_FORM_ID}-phone" class="input" placeholder="555-555-5555" />
-				</div>
-				<div class="form-group">
-					<label for="${ADD_FORM_ID}-street-address" class="input-label">Address</label>
-					<input type="text" name="streetAddress" id="${ADD_FORM_ID}-street-address" class="input" placeholder="Street Address" />
-					<label for="${ADD_FORM_ID}-address-locality" class="input-label required">City</label>
-					<input type="text" name="addressLocality" id="${ADD_FORM_ID}-address-locality" class="input" placeholder="Town" list="${ADD_FORM_ID}-towns-list" ${onChange}="${updateZip}" ${signalAttr}="${signal}" required="" />
-					<datalist id="${ADD_FORM_ID}-towns-list">
-						${TOWNS.map(town => `<option label="${town}" value="${town}"></option>`).join('\n')}
-					</datalist>
-					<label for="${ADD_FORM_ID}-postal-code" class="input-label required">Zip Code</label>
-					<input type="text" name="postalCode" id="${ADD_FORM_ID}-postal-code" class="input" pattern="\d{5}" inputmode="numeric" minlength="5" maxlength="5" placeholder="#####" list="${ADD_FORM_ID}-postal-list" required="" />
-					<datalist id="${ADD_FORM_ID}-postal-list">
-						${ZIPS.map(code => `<option value="${code}" label="${code}"></option>`).join('\n')}
-					</datalist>
-				</div>
-				<!--<div class="form-group">
-					<label for="${ADD_FORM_ID}-household-size" class="input-label required">How Many People Will This Feed?</label>
-					<input type="number" name="household" id="${ADD_FORM_ID}-household-size" class="input" placeholder="##" min="1" max="8" inputmode="numeric" required="" />
-				</div>-->
 				<div>
-					<p>Please provide the names for all of the people other than yourself this will be feeding</p>
-					<ol id="${HOUSEHOLD_LIST}" class="form-group"></ol>
-					<button type="button" class="btn btn-primary btn-lg" ${onClick}="${pantryAddHousehold}" ${signalAttr}="${sig}">
-						<svg xmlns="http://www.w3.org/2000/svg" fill="currentColor" width="12" height="16" viewBox="0 0 12 16" class="icon" role="presentation" aria-hidden="true">
-							<path fill-rule="evenodd" d="M12 9H7v5H5V9H0V7h5V2h2v5h5v2z"/>
-						</svg>
-						<span>Add Household Member</span>
-					</button>
-				</div>
-				<div class="form-group">
-					<label for="${ADD_FORM_ID}-date" class="input-label required">Pick a Date</label>
-					<input type="date" name="date" id="${ADD_FORM_ID}-date" class="input" ${attr({ value: new Date().toISOString().split('T')[0]})} required="" />
-				</div>
-				<div class="form-group">
-					<label for="${ADD_FORM_ID}-time" class="input-label required">Pick a Time</label>
-					<input type="time" name="time" id="${ADD_FORM_ID}-time" class="input" min="08:00" max="17:00" required="" />
+					<label for="${ADD_FORM_ID}-extra">Extra Trip</label>
+					<input type="checkbox" name="extraTrip" />
 				</div>
 				<div class="form-group">
 					<label for="${ADD_FORM_ID}-comments" class="input-label">
@@ -379,6 +359,20 @@ export default async ({ signal: sig }) => {
 	${getPantryHouseholdTemplate({ signal })}`;
 
 	frag.querySelector('details').append(video);
+
+	if (url.searchParams.size !== 0) {
+		whenLoaded({ signal: sig }).then(() => {
+			/**
+			 * @type {HTMLFormElement}
+			 */
+			const form = document.getElementById(ADD_FORM_ID);
+			const fields = form.elements;
+
+			url.searchParams.entries().forEach(([name, value]) => Promise.try(() => fields.namedItem(name).value = value));
+			document.getElementById(ADD_DIALOG_ID).showModal();
+			history.replaceState(history.state ?? {}, '', new URL(location.pathname, location.origin));
+		});
+	}
 
 	return frag;
 };